+++ projects/ngx-googlemaps-tracking-view/src/lib/components/center-marker/center-marker.ts
@@ -6,6 +6,7 @@ import { GoogleMapsWrapper } from '../../services/googlemaps-wrapper';
 import { HttpClient } from '@angular/common/http';
 
 let API_KEY = '';
+const INFOWINDOW_ID = 'center-pin';
 
 /**
  * A marker fixed to the center of the map which emits the geo-decoded location
@@ -23,23 +24,27 @@ let API_KEY = '';
  * <ng-template #infowindowLocation let-address>
  *   <div>
  *     <span *ngIf="!address">Loading...</span>
- *     <span *ngIf="!!address">{{address}}</span>
+ *       <div *ngIf="!!address">
+ *       {{address}}
+ *       <br>
+ *       <button (click)="map.centerPin.fix = !map.centerPin.fix">Fix</button>
+ *     </div>
  *   </div>
  * </ng-template>
  */
 export class CenterMarker {
   /**
-   * Emitted when the pin's location has changed (by zooming or dragging) and the map has become idle again.
+   * The current decoded address obtained from google. It becomes `null` while fetching.
    */
-  locationChanges = new EventEmitter<google.maps.LatLngLiteral>();
+  address: string;
   /**
    * Emitted when the the geocoded address has resolved.
    */
   addressChanges = new EventEmitter<string>();
   /**
-   * The current decoded address obtained from google. It becomes `null` while fetching.
+   * Emitted when the pin's location has changed (by zooming or dragging) and the map has become idle again.
    */
-  address: string;
+  locationChanges = new EventEmitter<google.maps.LatLngLiteral>();
 
   /** @ignore */
   private _infowindowTemplate: TemplateRef<any>
@@ -48,17 +53,40 @@ export class CenterMarker {
   /** @ignore */
   private _centerMarker: HTMLDivElement;
   /** @ignore */
+  private _fixedMarker: google.maps.Marker;
+  /** @ignore */
   private _mapEventSubscription = new Subscription();
   /** @ignore */
   private _geocoder = new google.maps.Geocoder();
   /** @ignore */
   private _idleListener: google.maps.MapsEventListener;
+  /** @ignore */
+  private _clickListener: google.maps.MapsEventListener;
 
 
   constructor(
     private googlemapsWrapper: GoogleMapsWrapper,
     private viewContainer,
     private http: HttpClient) { }
+
+  /**
+   * Fix the pin to its current position on map instead of floating over it.
+   * This value is reset on calling `disable()`.
+   */
+  get fix(): boolean { return !!this._fixedMarker; }
+  set fix(val: boolean) {
+    if (!this.googlemapsWrapper.isReady()) {
+      console.warn('Map is not ready');
+      return;
+    }
+
+    this.disable();
+    if (val) {
+      this.setFixedMarker();
+    } else {
+      this.enable(this._infowindowTemplate, API_KEY);
+    }
+  }
 
   /** @ignore */
   private get REVERSE_GEOCODING_URL() { return `https://maps.googleapis.com/maps/api/geocode/json?key=${API_KEY}&latlng=`; }
@@ -67,9 +95,11 @@ export class CenterMarker {
    * Activates the pin to show at the map's center
    * @param infowindowTemplate A template for rendering the infowindow on top of the marker.
    * It must have a single root element. The address string will be passed as implicit context.
-   * @param apiKey GoogleMaps API key
+   * @param apiKey GoogleMaps API key. TODO: remove the need for this param in future versions.
    */
   enable(infowindowTemplate: TemplateRef<any>, apiKey: string) {
+    this.removeFixedMarker();
+
     API_KEY = apiKey;
     this._infowindowTemplate = infowindowTemplate;
     if (this._centerMarker) return;
@@ -115,9 +145,9 @@ export class CenterMarker {
         })
     );
 
-    this.onMapIdle();
-    this.displayInfowindowOverPin();
+    this.updateAddress();
     this._idleListener = this.googlemapsWrapper.map.addListener('idle', () => this.onMapIdle());
+    this._clickListener = this.googlemapsWrapper.map.addListener('click', () => this.googlemapsWrapper.closeInfowindow(INFOWINDOW_ID));
   }
 
   /**
@@ -129,6 +159,9 @@ export class CenterMarker {
     this._mapEventSubscription.unsubscribe();
     this.googlemapsWrapper.closeInfowindow();
     google.maps.event.removeListener(this._idleListener);
+    google.maps.event.removeListener(this._clickListener);
+    this.removeFixedMarker();
+    this.googlemapsWrapper.closeInfowindow(INFOWINDOW_ID);
   }
 
   /** @ignore */
@@ -137,17 +170,45 @@ export class CenterMarker {
     const view = this.viewContainer.createEmbeddedView(this._infowindowTemplate, { $implicit: this.address });
     const content = view.rootNodes[0] as HTMLElement;
 
-    const center = this.googlemapsWrapper.map.getCenter();
-    const offset = this.googlemapsWrapper.getMetersPerPx(center.lat()) * 36;
-    const pos = google.maps.geometry.spherical.computeOffset(center, offset, 0);
+    const markerPosition = this.fix ? this._fixedMarker.getPosition() : this.googlemapsWrapper.map.getCenter();
+    const offset = this.googlemapsWrapper.getMetersPerPx(markerPosition.lat()) * 36;
+    const pos = google.maps.geometry.spherical.computeOffset(markerPosition, offset, 0);
 
-    this.googlemapsWrapper.openInfowindow(pos, content);
+    this.googlemapsWrapper.openInfowindow(pos, content, INFOWINDOW_ID)
+      .setOptions({ disableAutoPan: true });
+  }
+
+  /**
+   * Geodecode the current pin's location.
+   */
+  updateAddress() {
+    const markerPosition = this.fix ? this._fixedMarker.getPosition() : this.googlemapsWrapper.map.getCenter();
+    this._centerChanged$.next(markerPosition);
   }
 
   /** @ignore */
+  private setFixedMarker() {
+    if (this._fixedMarker) this.removeFixedMarker();
+    this._fixedMarker = new google.maps.Marker({
+      icon: 'https://maps.gstatic.com/mapfiles/markers/marker.png',
+      map: this.googlemapsWrapper.map,
+      position: this.googlemapsWrapper.map.getCenter(),
+    });
+
+    this._fixedMarker.addListener('click', () => this.displayInfowindowOverPin());
+  }
+
+  /** @ignore */
+  private removeFixedMarker() {
+    if (!this._fixedMarker) return;
+    this._fixedMarker.setMap(null);
+    google.maps.event.clearInstanceListeners(this._fixedMarker);
+    this._fixedMarker = null;
+  }
+
   private onMapIdle() {
     const center = this.googlemapsWrapper.map.getCenter();
     this.locationChanges.emit(center.toJSON());
-    this._centerChanged$.next(center);
+    this.updateAddress();
   }
 }