+++ projects/ngx-googlemaps-tracking-view/src/lib/services/googlemaps-wrapper.ts
@@ -233,6 +233,7 @@ export class GoogleMapsWrapper {
       infowindow.open(this.map, anchor);
     }
     infowindow['isOpen'] = true;
+    return infowindow;
   }
 
   /**